Fixes to join this branch to the sprint1
