+++ src/app/app.component.html
@@ -1,2 +1,2 @@
 <router-outlet></router-outlet>
-<app-map-box></app-map-box>
+<!--<app-map-box></app-map-box>-->
