+++ src/app/app.component.ts
@@ -5,7 +5,7 @@ import {Subscription} from 'rxjs';
 import {DomSanitizer} from '@angular/platform-browser';
 
 import {MatDialog, MatDialogRef, MAT_DIALOG_DATA} from '@angular/material';
-import {ChooseAudioCategoryComponent} from './choose-audio-category/choose-audio-category.component';
+import {ChooseAudioCategoryComponent} from './sections/record/components/choose-audio-category/choose-audio-category.component';
 
 @Component({
     selector: 'app-root',
@@ -15,22 +15,13 @@ import {ChooseAudioCategoryComponent} from './choose-audio-category/choose-audio
 export class AppComponent {
 
     constructor(private sanitizer: DomSanitizer,
-                private translate: TranslateService,
                 private audioRecord: AudioRecordService,
-                public dialog: MatDialog,
                 private translate: TranslateService) {
         translate.setDefaultLang('en');
     }
 
     switchLanguage(language: string) {
         this.translate.use(language);
-    }
-
-    openDialogChooseAudioCategory(): void {
-        const dialogRef = this.dialog.open(ChooseAudioCategoryComponent, {
-            width: '50%',
-            height: '40%',
-        });
     }
 
 }
