--- src/app/choose-audio-category/choose-audio-category.component.html
@@ -1,10 +1,0 @@
-<button mat-dialog-title id="boton" (click)="onNoClick()">Close</button>
-	<div id="modal" mat-dialog-content>
-	  <div id="title"><b>Choose Category</b></div>
-	  <div id="divMargin"><button (click)="clickTourism('tourism','experience','leisure')" id="tourism" class="buttonContent tourismButton">Tourism</button>
-	  <button id="experience" (click)="clickExperience('tourism','experience','leisure')" class="buttonContent experienceButton">Experience</button>
-    <button id="leisure" (click)="clickLeisure('tourism','experience', 'leisure')" class="buttonContent leisureButton">Leisure</button></div>
-    <br>
-    <div class="line"></div>
-    <div id="divMargin"><button (click)="saveCategory()" class="buttonContent saveButton">Save</button></div>
-	</div>
