--- src/app/choose-audio-category/choose-audio-category.component.scss
@@ -1,53 +1,0 @@
-#title{
-	  text-align: center;
-	  color: #8A3AB9; 
-	  font-size: 31px;
-}
-#boton{
-    float: right;
-    background-color: #ffffff;
-    border: none;
-    outline-color: transparent;
-    outline-style: none;
-    font-size: 24px;
-    color: #EB4339;
-    cursor: pointer;
-}
-#divMargin{
-    margin-top: 21px;
-    text-align: center;
-}
-.buttonContent{
-    border-radius: 10px;
-    margin: 10px;
-    cursor: pointer;
-    display: inline-block;
-    width: auto;
-    font-size: 22px;
-    height: 41px;
-    border: none;
-}
-
-.tourismButton{
-    background: #F24440;
-    color:#FFF;
-}
-
-.experienceButton{
-    background: #F57828;
-    color:#FFF;
-}
-
-.leisureButton{
-    background: #04B289;
-    color:#FFF;
-}
-
-.line{
-    border-bottom:solid 3px #8A3AB9;
-}
-
-.saveButton {
-    background: blue;
-    color:#FFF;
-}
