--- src/app/sections/record/record-routing.module.ts
@@ -1,14 +1,0 @@
-import {NgModule} from '@angular/core';
-import {Routes, RouterModule} from '@angular/router';
-import {AudioRecordComponent} from './components/audio-record/audio-record.component';
-
-const recordRoutes: Routes = [
-    {path: '/user', component: AudioRecordComponent},
-];
-
-@NgModule({
-    imports: [RouterModule.forRoot(recordRoutes)],
-    exports: [RouterModule]
-})
-export class AppRoutingModule {
-}
